refactor: route task persistence through data service
