+++ server/lib/events.ts
@@ -1,0 +1,5 @@
+import { EventEmitter } from "node:events";
+
+const events = new EventEmitter();
+
+export default events;
