+++ server/lib/sse.ts
@@ -1,4 +1,5 @@
 import type { Request, Response } from "express";
+import events from "./events.js";
 
 const clients: Response[] = [];
 
@@ -14,3 +15,5 @@ export function notifyClients(): void {
   const msg = "data: update\n\n";
   clients.forEach((res) => res.write(msg));
 }
+
+events.on("data:updated", notifyClients);
